humanoid-robortics/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere>=4.3.1
qdrant-client==1.9.1
google-generativeai==0.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
asyncpg==0.29.0
